Added table for displaying invoices.
Need to tweak the data for the Terms columns which initially had null values because of testing. Set the NULL values to empty strings and set the default value of column to an empty string (which still doesn't work). Eventually I had to set the Terms and InvoiceNumber into nullable strings. You may have to make adjustments depending on your database engine.

Added back button on the Create page.